deunify
This PR now streamlines some code and exports in generate so that stitchSchemas can follow functionality.
