+++ src/generate/index.ts
@@ -4,7 +4,6 @@ export { attachDirectiveResolvers } from './attachDirectiveResolvers';
 export { buildSchemaFromTypeDefinitions } from './buildSchemaFromTypeDefinitions';
 export { chainResolvers } from './chainResolvers';
 export { concatenateTypeDefs } from './concatenateTypeDefs';
-export { decorateWithLogger } from './decorateWithLogger';
 export {
   extractExtensionDefinitions,
   filterExtensionDefinitions,
