+++ src/index.ts
@@ -6,5 +6,6 @@ export * from './introspect';
 export * from './links';
 export * from './mock';
 export * from './scalars';
+export * from './stitch';
 export * from './wrap';
 export * from './utils';
