+++ src/stitch/index.ts
@@ -1,0 +1,1 @@
+export { stitchSchemas } from './stitchSchemas';
