+++ src/stitch/mergeInfo.ts
