+++ website/docs/schema-stitching.md
@@ -8,7 +8,7 @@ Schema stitching is the process of creating a single GraphQL schema from multipl
 
 One of the main benefits of GraphQL is that we can query all of our data as part of one schema, and get everything we need in one request. But as the schema grows, it might become cumbersome to manage it all as one codebase, and it starts to make sense to split it into different modules. We may also want to decompose your schema into separate microservices, which can be developed and deployed independently. We may also want to integrate our own schema with remote schemas.
 
-In these cases, we can use `makeExecutableSchema` to combine multiple existing GraphQL schemas together and produce a new schema that knows how to delegate parts of the query to the relevant subschemas. These subschemas can be either local to the server, or running on a remote server. They can even be services offered by 3rd parties, allowing us to connect to external data and create mashups.
+In these cases, we can use the `stitchSchemas` function to combine multiple existing GraphQL schemas together and produce a new schema that knows how to delegate parts of the query to the relevant subschemas. These subschemas can be either local to the server, or running on a remote server. They can even be services offered by 3rd parties, allowing us to connect to external data and create mashups.
 
 ## Basic example
 
@@ -18,6 +18,7 @@ In this example we'll stitch together two very simple schemas. In this case, we'
 import {
   makeExecutableSchema,
   addMocksToSchema,
+  stitchSchemas,
 } from 'graphql-tools';
 
 // Mocked chirp schema
@@ -56,7 +57,7 @@ const authorSchema = makeExecutableSchema({
 
 addMocksToSchema({ schema: authorSchema });
 
-export const schema = makeExecutableSchema({
+export const schema = stitchSchemas({
   subschemas: [
     { schema: chirpSchema, },
     { schema: authorSchema, },
@@ -101,7 +102,7 @@ const linkTypeDefs = `
 We can now merge these three schemas together:
 
 ```js
-export const schema = makeExecutableSchema({
+export const schema = stitchSchemas({
   subschemas: [
     { schema: chirpSchema, },
     { schema: authorSchema, },
@@ -116,14 +117,14 @@ We won't be able to query `User.chirps` or `Chirp.author` yet, however, because 
 
 How should these resolvers be implemented? When we resolve `User.chirps` or `Chirp.author`, we want to _delegate_ to the relevant root fields. To get from a user to the user's chirps, for example, we'll want to use the `id` of the user to call `Query.chirpsByAuthorId`. And to get from a chirp to its author, we can use the chirp's `authorId` field to call the existing `Query.userById` field.
 
-Resolvers can use the `delegateToSchema` function to forward parts of queries (or even whole new queries) to one of the subschemas that was passed to `makeExecutableSchema` (or even potentially other schemas).
+Resolvers can use the `delegateToSchema` function to forward parts of queries (or even whole new queries) to one of the subschemas that was passed to `stitchSchemas` (or even potentially other schemas).
 
 In order to delegate to these root fields, we'll need to make sure we've actually requested the `id` of the user or the `authorId` of the chirp. To avoid forcing users to add these fields to their queries manually, resolvers on a merged schema can define a `fragment` property that specifies the required fields, and they will be added to the query automatically.
 
 A complete implementation of schema stitching for these schemas might look like this:
 
 ```js
-const schema = makeExecutableSchema({
+const schema = stitchSchemas({
   subschemas: [
     { schema: chirpSchema, },
     { schema: authorSchema, },
@@ -219,12 +220,16 @@ We will now have a schema that has all fields and types prefixed with `Chirp_` a
 Now let's implement the resolvers:
 
 ```ts
+import {
+  stitchSchemas,
+} from 'graphql-tools';
+
 const chirpSubschema = {
   schema: chirpSchema,
   transforms: chirpSchemaTransforms,
 }
 
-export const schema = makeExecutableSchema({
+export const schema = stitchSchemas({
   subschemas: [
     chirpSubschema,
     { schema: authorSchema },
@@ -307,7 +312,7 @@ export type SubschemaConfig = {
   transforms?: Array<Transform>;
 };
 
-makeExecutableSchema({
+stitchSchemas({
   subschemas: Array<GraphQLSchema | SubschemaConfig>;
   typeDefs: string | DocumentNode;
   types: Array<GraphQLNamedType>;
@@ -398,9 +403,9 @@ type OnTypeConflict = (
 ) => GraphQLNamedType;
 ```
 
-The `onTypeConflict` option to `makeExecutableSchema` allows customization of type resolving logic.
+The `onTypeConflict` option to `stitchSchemas` allows customization of type resolving logic.
 
-The default behavior of `makeExecutableSchema` is to take the *last* encountered type of all the types with the same name, with a warning that type conflicts have been encountered. If specified, `onTypeConflict` enables explicit selection of the winning type.
+The default behavior of `stitchSchemas` is to take the *last* encountered type of all the types with the same name, with a warning that type conflicts have been encountered. If specified, `onTypeConflict` enables explicit selection of the winning type.
 
 For example, here's how we could select the *first* type among multiple types with the same name:
 
@@ -420,7 +425,7 @@ const onTypeConflict = (left, right, info) => {
 }
 ```
 
-When using schema transforms, `onTypeConflict` is often unnecessary, since transforms can be used to prevent conflicts before merging schemas. However, if you're not using schema transforms, `onTypeConflict` can be a quick way to make `makeExecutableSchema` produce more desirable results.
+When using schema transforms, `onTypeConflict` is often unnecessary, since transforms can be used to prevent conflicts before merging schemas. However, if you're not using schema transforms, `onTypeConflict` can be a quick way to make `stitchSchemas` produce more desirable results.
 
 #### inheritResolversFromInterfaces
